API: Simplify rotate API
Remove the many kwargs from the rotate function and replace with
one required argument: R, the rotation matrix. This makes the
code more readable. Users are now responsible for directly
using solve_rotation_ap or givens_rotation_matrix.

rotate is made part of the public API because it better reflects
the other functions such as union which can be used as
p1.union(p2) or union(p1, p2). rotation as a module function
in addition to the polytope class method is added for the same
reason.